Refactor: Extract map layout into a reusable component - Moved zoom controls, and search bar into separate MapLayout widget. - Updated SGW_MAP_Page to use the new MapLayout widget.
