+++ concordia_nav/lib/core/ui/widgets/sgw_map_view.dart
@@ -1,7 +1,7 @@
 import 'package:concordia_nav/core/ui/widgets/search_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:concordia_nav/core/ui/widgets/custom_appbar.dart';
-import 'package:concordia_nav/core/ui/widgets/map_controls.dart';
+import 'package:concordia_nav/core/ui/widgets/map_layout.dart';
 
 class SgwMapPage extends StatefulWidget {
   const SgwMapPage({super.key});
@@ -15,25 +15,12 @@ class SgwMapPageState extends State<SgwMapPage> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: customAppBar(context, 'SGW Map'),
-      body: Stack(
-        children: [
-          Placeholder(),
-          Positioned(
-            top: 10,
-            left: 15,
-            right: 15,
-            child: SearchBarWidget(controller: TextEditingController(),) ,
-          ),
-          Positioned(
-            top: 100,
-            right: 15,
-            child: MapControls(
-              onMyLocation: () {}, // Replace with your my location function
-              onZoomIn: () {}, // Replace with your zoom in function
-              onZoomOut: () {}, // Replace with your zoom out function
-            ),
-          ),
-      ])
+      body: MapLayout(
+        searchController: TextEditingController(),
+        onMyLocation: () {},
+        onZoomIn: () {},
+        onZoomOut: () {}
+      )
     );
   }
 }
